refactor: update domain model data types for precision and scalability
- Se refactorizó la entidad Product cambiando id a Long para mayor escalabilidad.
- Se reemplazó Double por BigDecimal en price para garantizar precisión financiera.
- Se actualizó ProductRepository para soportar el nuevo tipo de identificador.
- Se ajustaron temporalmente las firmas en ProductController para evitar errores de compilación.
- Se migró la aplicación a un esquema de base de datos independiente (api_inventario).

+++ src/main/java/io/github/torres/controller/ProductController.java
@@ -35,14 +35,14 @@ public class ProductController {
 
     // 3. DELETE: Remove a product by ID
     @DeleteMapping("/{id}")
-    public void deleteProduct(@PathVariable Integer id) {
+    public void deleteProduct(@PathVariable Long id) {
         // @PathVariable extracts the ID from the URL (e.g., /api/products/5)
         productRepository.deleteById(id);
     }
 
     // 4. PUT: Update an existing product
     @PutMapping("/{id}")
-    public Product updateProduct(@PathVariable Integer id, @RequestBody Product product) {
+    public Product updateProduct(@PathVariable Long id, @RequestBody Product product) {
 
         if (product != null && id > 0) {
             // Ratrieve the original product from the database (throws an exception if not
